# Backendbooking/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Space Booking API
  version: 1.0.0
  description: |
    Kontrak 50 operasi UKK Paket B. Semua data bisnis tenant-scoped.
    x-maker-key dan x-app-key adalah alias; jika keduanya hadir nilainya harus sama.
    Kontrak ini sengaja memperketat upload menjadi bearer-only, melindungi maker/list,
    menggunakan QR opaque tanpa app key, dan mempertahankan body array collection dengan
    metadata pagination pada X-Page/X-Per-Page/X-Total-Count. Lihat api-contract.md untuk conflict log.
servers:
  - url: https://learn.smktelkom-mlg.sch.id/coworking
    description: Base URL contoh sumber UKK
  - url: http://localhost:3000
    description: Development lokal
...
paths:
...
  /api/auth/register/member:
    post:
      tags: [Auth]
      operationId: registerMember
      summary: Registrasi member dalam tenant
      security:
        - makerKey: []
        - appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterMemberDto' }
      responses:
        '201':
          description: Member dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterMemberAuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/register/admin-space:
    post:
      tags: [Auth]
      operationId: registerAdminSpace
      summary: Registrasi admin dan profil coworking
      security:
        - makerKey: []
        - appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterAdminSpaceDto' }
      responses:
        '201':
          description: Admin dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterAdminAuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Login member atau admin dalam tenant
      security:
        - makerKey: []
        - appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginDto' }
      responses:
        '200':
          description: Login berhasil; profil role terkait dapat null/absent sesuai data sumber dan toleransi frontend
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginUserAuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/profile:
    get:
      tags: [Auth]
      operationId: getCurrentUserProfile
      summary: Profil user login
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Profil user; hanya profil role terkait yang wajib, profil lain dapat null/absent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfileResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/reservasi:
    post:
      tags: [Reservations]
      operationId: createReservation
      summary: Membuat reservasi member secara atomik
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReservationDto' }
      responses:
        '201':
          description: Reservasi dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/admin/profile:
    get:
      tags: [Admin Profile]
      operationId: getAdminCoworkingProfile
      summary: Profil coworking admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Profil owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OwnerResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [Admin Profile]
      operationId: updateAdminCoworkingProfile
      summary: Update profil coworking admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCoworkingProfileDto' }
      responses:
        '200':
          description: Profil diperbarui
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OwnerResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/members:
    get:
      tags: [Admin Members]
      operationId: listAdminMembers
      summary: Daftar member tenant
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - name: search
          in: query
          schema: { type: string, minLength: 1, maxLength: 100 }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Daftar member
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Admin Members]
      operationId: createAdminMember
      summary: Tambah member oleh admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMemberAdminDto' }
      responses:
        '201':
          description: Member dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/members/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [Admin Members]
      operationId: getAdminMemberById
      summary: Detail member
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Detail member
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [Admin Members]
      operationId: updateAdminMember
      summary: Update member
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMemberAdminDto' }
      responses:
        '200':
          description: Member diperbarui
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Admin Members]
      operationId: deleteAdminMember
      summary: Arsipkan/hapus member tanpa merusak histori
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Member diarsipkan atau dihapus
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/admin/reservasi:
    get:
      tags: [Admin Reservations]
      operationId: listAdminReservations
      summary: Daftar dan filter reservasi owner
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/Month' }
        - { $ref: '#/components/parameters/Year' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ReservationStatus' }
        - name: id_space
          in: query
          schema: { $ref: '#/components/schemas/PositiveId' }
        - name: tanggal
          in: query
          schema: { $ref: '#/components/schemas/DateOnly' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Reservasi owner
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminReservationListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/reservasi/{id}/status:
    patch:
      tags: [Admin Reservations]
      operationId: updateAdminReservationStatus
      summary: Konfirmasi atau batalkan reservasi
      description: Enum sumber dipertahankan, tetapi aktif dan selesai hanya dicapai melalui endpoint check-in/check-out.
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateReservationStatusDto' }
      responses:
        '200':
          description: Status diperbarui
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationStatusResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/reservasi/{id}/check-in:
    post:
      tags: [Admin Reservations]
      operationId: checkInAdminReservation
      summary: Check-in reservasi disetujui setelah inspeksi e-ticket EP-022
      description: Inspeksi tiket mendahului operasi ini, tetapi authorization, ownership, dan state tetap diverifikasi ulang server-side.
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Check-in berhasil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckInResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/reservasi/{id}/check-out:
    post:
      tags: [Admin Reservations]
      operationId: checkOutAdminReservation
      summary: Check-out reservasi aktif
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Check-out berhasil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckOutResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/reports/monthly:
    get:
      tags: [Reports]
      operationId: getAdminMonthlyReport
      summary: Rekap bulanan dan distribusi per tipe
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/Month' }
        - { $ref: '#/components/parameters/Year' }
      responses:
        '200':
          description: Laporan bulanan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthlyReportResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/reports/income:
    get:
      tags: [Reports]
      operationId: getAdminIncomeReport
      summary: Alias ringkas pendapatan bulanan
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/Month' }
        - { $ref: '#/components/parameters/Year' }
      responses:
        '200':
          description: Pendapatan bersih bulanan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncomeReportResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT App Maker, member, admin_space, atau Guru/Penguji sesuai endpoint.
    makerKey:
      type: apiKey
      in: header
      name: x-maker-key
      description: Tenant key utama.
    appKey:
      type: apiKey
      in: header
      name: x-app-key
      description: Alias kompatibilitas untuk x-maker-key; bila keduanya ada harus sama.
  headers:
    XPage:
      description: Nomor halaman efektif (1-based).
      schema: { type: integer, minimum: 1 }
    XPerPage:
      description: Batas efektif item per halaman.
      schema: { type: integer, minimum: 1, maximum: 100 }
    XTotalCount:
      description: Jumlah total item yang cocok sebelum pagination.
      schema: { type: integer, minimum: 0 }
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/PositiveId' }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Month:
      name: month
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 12 }
    Year:
      name: year
      in: query
      required: false
      schema: { type: integer, minimum: 2000, maximum: 2100 }
  responses:
    BadRequest:
      description: Input/header tidak valid atau aturan bisnis dasar gagal
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { status: false, statusCode: 400, message: Permintaan tidak valid, error: Bad Request, timestamp: '2026-08-27T08:00:00.000Z' }
    Unauthorized:
      description: Credential hilang atau invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Role/ownership/state tidak mengizinkan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource tidak ada atau disamarkan karena scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Duplicate, overlap, stale write, atau transisi ilegal
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    TooManyRequests:
      description: Rate limit terlampaui
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
          description: Detik sebelum mencoba kembali
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Kegagalan internal tanpa stack/SQL/secret
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    PositiveId:
      type: integer
      format: int64
      minimum: 1
      example: 1
...
    DateOnly:
      type: string
      format: date
      example: '2026-08-30'
    TimeHHmm:
      type: string
      pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
      example: '09:00'
...
    UserRole:
      type: string
      enum: [member, admin_space]
    ReservationStatus:
      type: string
      enum: [belum_dikonfirm, disetujui, aktif, selesai, dibatalkan]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [status, statusCode, message, error, timestamp]
      properties:
        status: { type: boolean, enum: [false] }
        statusCode: { type: integer, enum: [400, 401, 403, 404, 409, 413, 429, 500] }
        message: { type: string }
        error: { type: string }
        timestamp: { type: string, format: date-time }
...
    SuccessBase:
      type: object
      required: [status, statusCode, message, timestamp]
      properties:
        status: { type: boolean, enum: [true] }
        statusCode: { type: integer, enum: [200, 201] }
        message: { type: string }
        timestamp: { type: string, format: date-time }
...
    RegisterMemberDto:
      type: object
      additionalProperties: false
      required: [username, password, nama_member, instansi, alamat, telp]
      properties:
        username: { type: string, minLength: 1, maxLength: 100, example: johndoe }
        password: { type: string, format: password, minLength: 6, maxLength: 128, writeOnly: true, example: Secret123! }
        nama_member: { type: string, minLength: 1, maxLength: 200, example: John Doe }
        instansi: { type: string, minLength: 1, maxLength: 200, example: Universitas Indonesia }
        alamat: { type: string, minLength: 1, maxLength: 1000, example: Jl. Sudirman No. 123 }
        telp: { type: string, minLength: 3, maxLength: 32, example: '081234567890' }
        foto: { type: string, minLength: 1, maxLength: 512, nullable: true, example: member_john.jpg }
    RegisterAdminSpaceDto:
      type: object
      additionalProperties: false
      required: [username, password, nama_coworking, nama_pemilik, telp]
      properties:
        username: { type: string, minLength: 1, maxLength: 100, example: admin_space1 }
        password: { type: string, format: password, minLength: 6, maxLength: 128, writeOnly: true, example: Admin123! }
        nama_coworking: { type: string, minLength: 1, maxLength: 200, example: Moklet Hub Coworking }
        nama_pemilik: { type: string, minLength: 1, maxLength: 200, example: Ahmad Bidin }
        telp: { type: string, minLength: 3, maxLength: 32, example: '081298765432' }
        alamat: { type: string, minLength: 1, maxLength: 1000, nullable: true, description: Ekstensi kebutuhan profil. }
        deskripsi_fasilitas: { type: string, minLength: 1, maxLength: 4000, nullable: true, description: Ekstensi kebutuhan profil. }
    LoginDto:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 100, example: johndoe }
        password: { type: string, format: password, minLength: 1, maxLength: 128, writeOnly: true, example: Secret123! }
...
    Owner:
      type: object
      additionalProperties: false
      required: [id, nama_coworking, nama_pemilik, telp]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        nama_coworking: { type: string }
        nama_pemilik: { type: string }
        telp: { type: string }
        alamat: { type: string, nullable: true }
        deskripsi_fasilitas: { type: string, nullable: true }
    UserAuthBase:
      type: object
      additionalProperties: false
      required: [id, username, role, access_token]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        username: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        maker_id:
          allOf: [{ $ref: '#/components/schemas/PositiveId' }]
          description: Source-compatible optional field; frontend must tolerate absence.
        member: { type: object, allOf: [{ $ref: '#/components/schemas/Member' }], nullable: true }
        space_owner: { type: object, allOf: [{ $ref: '#/components/schemas/Owner' }], nullable: true }
        access_token: { type: string, readOnly: true }
    RegisterMemberAuth:
      allOf:
        - { $ref: '#/components/schemas/UserAuthBase' }
        - type: object
          required: [member]
          properties:
            role: { type: string, enum: [member] }
            member: { $ref: '#/components/schemas/Member' }
    RegisterAdminAuth:
      allOf:
        - { $ref: '#/components/schemas/UserAuthBase' }
        - type: object
          required: [space_owner]
          properties:
            role: { type: string, enum: [admin_space] }
            space_owner: { $ref: '#/components/schemas/Owner' }
    RegisterMemberAuthResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/RegisterMemberAuth' }
    RegisterAdminAuthResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/RegisterAdminAuth' }
    LoginUserAuthResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/UserAuthBase' }
...
    UserProfileResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/UserProfile' }
...
    CreateReservationDto:
      type: object
      additionalProperties: false
      required: [id_space, tanggal_reservasi, jam_mulai, durasi_jam]
      properties:
        id_space: { $ref: '#/components/schemas/PositiveId' }
        tanggal_reservasi: { $ref: '#/components/schemas/DateOnly' }
        jam_mulai: { $ref: '#/components/schemas/TimeHHmm' }
        durasi_jam: { type: integer, minimum: 1, example: 3 }
        id_diskon: { type: integer, format: int64, minimum: 1, nullable: true }
        kode_promo: { type: string, minLength: 1, maxLength: 100, nullable: true, example: DISKONHEMAT20 }
...
    ReservationResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/Reservation' }
...
    ReservationStatusResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ReservationStatusData' }
    OwnerResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/Owner' }
    UpdateCoworkingProfileDto:
      type: object
      additionalProperties: false
      required: [nama_coworking, nama_pemilik, telp]
      properties:
        nama_coworking: { type: string, minLength: 1, maxLength: 200 }
        nama_pemilik: { type: string, minLength: 1, maxLength: 200 }
        telp: { type: string, minLength: 3, maxLength: 32 }
        alamat: { type: string, minLength: 1, maxLength: 1000, nullable: true, description: Ekstensi backward-compatible. }
        deskripsi_fasilitas: { type: string, minLength: 1, maxLength: 4000, nullable: true, description: Ekstensi backward-compatible. }
    MemberResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/Member' }
    MemberListResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/Member' }
    CreateMemberAdminDto:
      allOf:
        - { $ref: '#/components/schemas/RegisterMemberDto' }
    UpdateMemberAdminDto:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nama_member: { type: string, minLength: 1, maxLength: 200 }
        instansi: { type: string, minLength: 1, maxLength: 200 }
        alamat: { type: string, minLength: 1, maxLength: 1000 }
        telp: { type: string, minLength: 3, maxLength: 32 }
        password: { type: string, format: password, minLength: 6, maxLength: 128, writeOnly: true }
        foto: { type: string, minLength: 1, maxLength: 512, nullable: true }
...
    DeleteResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/DeleteData' }
...
    AdminReservationListResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/AdminReservation' }
    UpdateReservationStatusDto:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/ReservationStatus'
      description: Aktif/selesai dipertahankan pada enum sumber tetapi ditolak di endpoint ini; gunakan check-in/check-out.
...
    CheckInResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/CheckInData' }
...
    CheckOutResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/CheckOutData' }
...
    MonthlyReportResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/MonthlyReport' }
...
    IncomeReportResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/IncomeReport' }
...